requirements.txt
anthropic>=0.49.0
google-generativeai>=0.8.0
PyGithub>=2.3.0
duckduckgo-search>=6.2.0
requests>=2.32.0
python-dotenv>=1.0.0
